refactor: extract notification logic into reusable mixin
