+++ packages/notification/package.json
@@ -37,6 +37,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/component-base": "24.6.0-alpha4",
     "@vaadin/lit-renderer": "24.6.0-alpha4",
